+++ android/src/main/java/com/bespot/gatekeepersdkcapacitorplugin/SafeSDK.kt
@@ -1,0 +1,59 @@
+package com.bespot.gatekeepersdkcapacitorplugin
+
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
+import com.bespot.shared.common.models.Action
+import com.bespot.shared.core.Failure
+import com.bespot.shared.core.FraudulentCheckObserver
+import com.bespot.shared.core.SafeSdk
+import com.getcapacitor.Logger
+
+class SafeSDK: FraudulentCheckObserver {
+
+    private val safeSdk: SafeSdk = SafeSdk
+
+    private val _checkResult: MutableState<Result<CheckResult>> = mutableStateOf(Result.success(CheckResult.empty()))
+
+    init {
+        //safeSdk.logging(true) // Uncomment if you want to see logs during the app launch
+    }
+
+    override fun onSuccess(
+        action: Action,
+        signature: String,
+    ) {
+        _checkResult.value = Result.success(CheckResult(action = action, ticket = signature))
+    }
+
+    override fun onError(error: Failure) {
+        _checkResult.value = Result.failure(error)
+    }
+
+    fun echo(value: String?): String? {
+        Logger.info("Echo", value)
+        safeSdk.check(this)
+        return value
+    }
+
+    fun check(callback: (result: Result<CheckResult>) -> Unit) {
+        safeSdk.check(object : FraudulentCheckObserver {
+            override fun onSuccess(action: Action, signature: String) {
+                // Send successful result back through callback
+                callback(Result.success(CheckResult(action = action, ticket = signature)))
+            }
+
+            override fun onError(error: Failure) {
+                // Send error back through callback
+                callback(Result.failure(error))
+            }
+        })
+    }
+
+
+    fun setUserId(value: String) {
+        safeSdk.setUserID(value)
+    }
+    fun enableLogging(value: Boolean) {
+        safeSdk.logging(value)
+    }
+}