+++ example-app/android/app/src/main/java/com/bespot/capacitorplugin/androidapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.plugin;
+package com.bespot.capacitorplugin.androidapp;
 
 import com.getcapacitor.BridgeActivity;