+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v1.1.0 (2025-12-16)
+
+### Added
+- Plugin for Android
+- `SafeSDKError` interface 
+
+### Changed
+- `askForPermissions()` renamed to `askForLocationPermissions()`
+- `ActionType` type
+
 # v1.0.0 (2025-12-05)
 
 ### Added